fixed an albedo miscalculation problem

+++ srcs/material.c
@@ -6,13 +6,11 @@
 /*   By: mclaudel <mclaudel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/12/03 15:52:20 by mclaudel          #+#    #+#             */
-/*   Updated: 2020/01/13 14:31:33 by mclaudel         ###   ########.fr       */
+/*   Updated: 2020/01/21 17:51:24 by mclaudel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <material.h>
-
-#include <stdio.h>
 
 t_material		*create_material(
 	unsigned int color,
@@ -26,9 +24,9 @@ t_material		*create_material(
 	if (m == NULL)
 		return (NULL);
 	c.v = color;
-	m->albedo.x = c.color.r / 0xff * (1 - metallic);
-	m->albedo.y = c.color.g / 0xff * (1 - metallic);
-	m->albedo.z = c.color.b / 0xff * (1 - metallic);
+	m->albedo.x = (double) c.color.r / 0xff * (1 - metallic);
+	m->albedo.y = (double) c.color.g / 0xff * (1 - metallic);
+	m->albedo.z = (double) c.color.b / 0xff * (1 - metallic);
 	m->refractive = 0;
 	m->refractionindex = 0;
 	m->specular = 1 - roughness;
